REFACTOR : refactored code, added documentation to methods, parsing logic is in distinct project to be modular
