+++ Logic/MainLogic.cs
@@ -1,10 +1,6 @@
-﻿using HydrothermalJunctionDetector.UI;
-using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using HydrothermalVentFileParser.UI;
+using HydrothermalVentFileParser;
+
 
 namespace HydrothermalJunctionDetector.Logic
 {
