+++ UI/ConsoleLogger.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using HydrothermalVentFileParser.Logging;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -6,7 +7,7 @@ using System.Threading.Tasks;
 
 namespace HydrothermalJunctionDetector.UI
 {
-    internal class ConsoleLogger : ILogger
+    public class ConsoleLogger : ILogger
     {
         public void LogWriteLine(string message)
         {
@@ -20,7 +21,7 @@ namespace HydrothermalJunctionDetector.UI
 
         public void LogError(string message)
         {
-            Console.Write(message);
+            Console.WriteLine(message);
         }
     }
 }
